requirements.txt
Django==2.2.7
django-user-agents==0.4.0
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
ua-parser==0.10.0
user-agents==2.1
